+++ frontends/benchmarks/extraction/invalid/ValOverrideAbstract.scala
@@ -1,0 +1,9 @@
+object ValOverrideAbstract {
+  sealed abstract class AAA {
+    def f: BigInt
+  }
+
+  case class BBB() extends AAA {
+    val f: BigInt = 0
+  }
+}